backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
celery[redis]==5.4.0
redis==5.0.8
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
pymupdf==1.24.11
Pillow==10.4.0
opencv-python-headless==4.10.0.84
deskew==1.5.1
anthropic==0.36.0
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
boto3==1.35.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.0
pydantic-settings==2.5.2
pytest==8.3.0
pytest-asyncio==0.24.0
